requirements.txt
python-telegram-bot
vk_api
bs4
